backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.34
pymysql==1.1.1
python-dotenv==1.0.1
pydantic==2.9.2
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
aiofiles==24.1.0
numpy==2.1.2
PyJWT==2.9.0
loguru==0.7.2
faiss-cpu==1.8.0
fastembed==0.4.1